♻️ refactor card component
