+++ nativeApp/src/lib/ui/CardWithButtons/Buttons.tsx
@@ -1,36 +1,20 @@
 import React, { FC } from 'react';
 import styled from 'styled-components/native';
 import { MaterialCommunityIcons, MaterialIcons } from '@expo/vector-icons';
-import { Card as C, Button as B } from 'native-base';
+import { Button as B } from 'native-base';
 
-import { ColoursEnum } from './colours';
+import { ColoursEnum } from '../colours';
 
 /*
  * Types
  */
 
-export type RemoveType = 'archive' | 'delete' | 'undo';
 type Props = {
-  removeType: RemoveType;
 }
 
 /*
  * Styles
  */
-const Card = styled(C)`
-  width: 85%;
-  marginBottom: 10;
-`;
-
-const TopContainer = styled.View`
-  marginLeft: 10;
-  marginTop: 10;
-  marginRight: 10;
-  paddingBottom: 5;
-  borderBottomWidth: 1;
-  borderColor: ${ColoursEnum.grey};
-`;
-
 const ButtonContainer = styled.View`
   flexDirection: row;
 `;
@@ -57,7 +41,10 @@ const BorderRight = styled.View`
   borderColor: ${ColoursEnum.grey};
 `;
 
-const ArchiveButton: FC = () => (
+/*
+ * Components
+ */
+export const ArchiveButton: FC = () => (
   <ButtonContainer>
     <Button transparent>
       <BorderRight>
@@ -72,7 +59,7 @@ const ArchiveButton: FC = () => (
   </ButtonContainer>
 );
 
-const DeleteButton: FC = () => (
+export const DeleteButton: FC = () => (
   <ButtonContainer>
     <Button transparent>
       <BorderRight>
@@ -87,36 +74,12 @@ const DeleteButton: FC = () => (
   </ButtonContainer>
 );
 
-const UndoButton: FC = () => (
+export const UndoButton: FC = () => (
   <ButtonContainer>
-    <Button transparent>
-      {/* <BorderRight>
-        <MaterialCommunityIcons name="square-edit-outline" outline size={20} color={ColoursEnum.purple} />
-        <LinkText>Edit</LinkText>
-      </BorderRight> */}
-    </Button>
+    <Button transparent />
     <Button transparent>
       <MaterialCommunityIcons name="undo" outline size={20} color={ColoursEnum.purple} />
       <LinkText>Unarchive</LinkText>
     </Button>
   </ButtonContainer>
 );
-/*
- * Component
- */
-const CardWithButtons: FC<Props> = (props) => {
-  const { children, removeType } = props;
-  return (
-    <Card>
-      <TopContainer>
-        {children}
-      </TopContainer>
-      {removeType === 'archive' && <ArchiveButton />}
-      {removeType === 'delete' && <DeleteButton />}
-      {removeType === 'undo' && <UndoButton />}
-    </Card>
-
-  );
-};
-
-export default CardWithButtons;
